refactor(vscode): the timeout defaults and the process-tree kill each live in one place
